Added boiler-plate code for IPDetails

+++ Models/ReportDTO.cs
@@ -3,6 +3,6 @@ namespace IP2C_Web_API.Models;
 public class ReportDTO
 {
     public string? CountryName { get; set; }
-    public long? AddressesCount { get; set; }
+    public int? AddressesCount { get; set; }
     public DateTime? LastAddressUpdated { get; set; }
 }
